Create a new portion when product is selected
