+++ client/app/styles/forms/f-portion.styl
@@ -4,3 +4,8 @@
     @media $max767
       font-size $h2-font-size
       margin-bottom 10px
+
+  &__subtle
+    font-size $h3-font-size
+    margin-bottom 10px
+    text-align center
