+++ client/app/templates/components/b-product.hbs
@@ -5,5 +5,5 @@
   <div class="b-product__cost">{{format-currency product.cost}}</div>
 </div>
 <div class="b-product__control">
-  <button class="button button_submit">Добавить</button>
+  <button class="button button_submit" {{action 'add'}}>Добавить</button>
 </div>
