+++ client/app/templates/components/f-portion.hbs
@@ -6,10 +6,14 @@
     {{b-order-description order=order}}
   </div>
 
+  <h3 class="f-portion__subtle">
+    Выберите любые понравившиеся блюда, чтобы добавить их к заказу
+  </h3>
+
   <ul class="b-list b-product-list">
     {{#each products as |product|}}
       <li class="b-list__item">
-        {{b-product product=product}}
+        {{b-product product=product add=(action 'addToOrder' order)}}
       </li>
     {{/each}}
   </ul>
