+++ client/app/templates/new-portion-order.hbs
@@ -1,1 +1,1 @@
-{{f-portion order=model products=products submit=(action 'addToOrder' model)}}
+{{f-portion order=model products=products}}
